test: move test to dedicated fixture
