+++ tests/fixtures/after/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "after",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "postinstall": "next build",
+    "dev": "next dev",
+    "build": "next build"
+  },
+  "dependencies": {
+    "next": "latest",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
+  },
+  "test": {
+    "dependencies": {
+      "next": ">=15.0.0"
+    }
+  }
+}
