+++ tests/fixtures/simple/next.config.js
@@ -8,7 +8,6 @@ const nextConfig = {
     outputFileTracingIncludes: {
       '/': ['public/**'],
     },
-    after: true,
   },
   images: {
     remotePatterns: [
